+++ src/NetStandard/E.Standard.WebGIS.CmsSchema/ArcServerTileService.cs
@@ -71,7 +71,7 @@ public class ArcServerTileService : CopyableNode, IAuthentification, ICreatable,
     }
 
     [DisplayName("Rendering")]
-    [Description("Für Luftbilder 'Quality' verwenden. Für Ortspläne (mit Text) 'Readablility'...")]
+    [Description("Für Luftbilder 'Quality' verwenden. Für Ortspläne (mit Text) 'Readablility'. Für MapServices mit Layern und Maßstabsgrenzen 'ScaleDependentLayers'")]
     public TileGridRendering Rendering
     {
         get { return _rendering; }