+++ src/NetStandard/E.Standard.WebGIS.CmsSchema/WMTSService.cs
@@ -85,7 +85,7 @@ public class WMTSService : CopyableNode, IAuthentification, ICreatable, IEditabl
     }
 
     [DisplayName("Rendering")]
-    [Description("Für Luftbilder 'Quality' verwenden. Für Ortspläne (mit Text) 'Readablility'...")]
+    [Description("Für Luftbilder 'Quality' verwenden. Für Ortspläne (mit Text) 'Readablility'. Für MapServices mit Layern und Maßstabsgrenzen 'ScaleDependentLayers'")]
     public TileGridRendering Rendering
     {
         get { return _rendering; }